requirements.txt
discord.py
asyncio
python-dotenv
